Aplicando un refactor a la función show-run de arena
