+++ frontend/www/js/omegaup/components/problem/Details.vue
@@ -250,8 +250,7 @@
 
 <script lang="ts">
 import { Vue, Component, Prop, Ref, Emit, Watch } from 'vue-property-decorator';
-import { omegaup } from '../../omegaup';
-import { messages, types } from '../../api_types';
+import { types } from '../../api_types';
 import T from '../../lang';
 import * as time from '../../time';
 import * as ui from '../../ui';
@@ -301,32 +300,6 @@ export enum PopupDisplayed {
   Reviewer,
 }
 
-const numericSort = <T extends { [key: string]: any }>(key: string) => {
-  const isDigit = (ch: string) => '0' <= ch && ch <= '9';
-  return (x: T, y: T) => {
-    let i = 0,
-      j = 0;
-    for (; i < x[key].length && j < y[key].length; i++, j++) {
-      if (isDigit(x[key][i]) && isDigit(x[key][j])) {
-        let nx = 0,
-          ny = 0;
-        while (i < x[key].length && isDigit(x[key][i]))
-          nx = nx * 10 + parseInt(x[key][i++]);
-        while (j < y[key].length && isDigit(y[key][j]))
-          ny = ny * 10 + parseInt(y[key][j++]);
-        i--;
-        j--;
-        if (nx != ny) return nx - ny;
-      } else if (x[key][i] < y[key][j]) {
-        return -1;
-      } else if (x[key][i] > y[key][j]) {
-        return 1;
-      }
-    }
-    return x[key].length - i - (y[key].length - j);
-  };
-};
-
 @Component({
   components: {
     FontAwesomeIcon,
@@ -374,7 +347,8 @@ export default class ProblemDetails extends Vue {
   @Prop() selectedPrivateTags!: string[];
   @Prop() hasBeenNominated!: boolean;
   @Prop({ default: null }) runDetailsData!: types.RunDetails | null;
-  @Prop() guid!: string;
+  @Prop({ default: null }) guid!: string;
+  @Prop({ default: null }) problemAlias!: string;
   @Prop() isAdmin!: boolean;
   @Prop({ default: false }) showVisibilityIndicators!: boolean;
   @Prop({ default: false }) shouldShowTabs!: boolean;
@@ -465,7 +439,14 @@ export default class ProblemDetails extends Vue {
   }
 
   onRunDetails(guid: string): void {
-    this.$emit('show-run', this, guid);
+    this.$emit('show-run', {
+      request: {
+        guid,
+        isAdmin: this.isAdmin,
+        problemAlias: this.problem.alias,
+      },
+      target: this,
+    });
     this.currentPopupDisplayed = PopupDisplayed.RunDetails;
   }
 
@@ -596,55 +577,6 @@ export default class ProblemDetails extends Vue {
     });
   }
 
-  displayRunDetails(guid: string, data: messages.RunDetailsResponse): void {
-    let sourceHTML,
-      sourceLink = false;
-    if (data.source?.indexOf('data:') === 0) {
-      sourceLink = true;
-      sourceHTML = data.source;
-    } else if (data.source == 'lockdownDetailsDisabled') {
-      sourceHTML =
-        (typeof sessionStorage !== 'undefined' &&
-          sessionStorage.getItem(`run:${guid}`)) ||
-        T.lockdownDetailsDisabled;
-    } else {
-      sourceHTML = data.source;
-    }
-
-    const detailsGroups = data.details && data.details.groups;
-    let groups = undefined;
-    if (detailsGroups && detailsGroups.length) {
-      detailsGroups.sort(numericSort('group'));
-      for (const detailGroup of detailsGroups) {
-        if (!detailGroup.cases) {
-          continue;
-        }
-        detailGroup.cases.sort(numericSort('name'));
-      }
-      groups = detailsGroups;
-    }
-
-    Vue.set(
-      this,
-      'currentRunDetailsData',
-      Object.assign({}, data, {
-        logs: data.logs || '',
-        judged_by: data.judged_by || '',
-        source: sourceHTML,
-        source_link: sourceLink,
-        source_url: window.URL.createObjectURL(
-          new Blob([data.source || ''], { type: 'text/plain' }),
-        ),
-        source_name: `Main.${data.language}`,
-        groups: groups,
-        show_diff: this.isAdmin ? data.show_diff : 'none',
-        feedback: omegaup.SubmissionFeedback.None as omegaup.SubmissionFeedback,
-      }),
-    );
-
-    this.$emit('change-show-run-location', { guid });
-  }
-
   @Emit('update:activeTab')
   onTabSelected(tabName: string): string {
     if (this.selectedTab === 'clarifications') {
@@ -685,7 +617,14 @@ export default class ProblemDetails extends Vue {
   @Watch('shouldShowRunDetails')
   onShouldShowRunDetailsChanged(newValue: boolean): void {
     if (newValue && this.guid) {
-      this.$emit('show-run', this, this.guid);
+      this.$emit('show-run', {
+        request: {
+          guid: this.guid,
+          isAdmin: this.isAdmin,
+          problemAlias: this.currentRunDetailsData?.alias,
+        },
+        target: this,
+      });
     }
   }
 }
